exception in service and repo

See merge request sep-ws-2021-2022/sep-ws21g04/lases!225

+++ app/lases/src/main/java/de/lases/control/backing/ScientificForumBacking.java
@@ -10,6 +10,9 @@ import de.lases.control.internal.Pagination;
 import de.lases.control.internal.SessionInformation;
 import de.lases.global.transport.*;
 import jakarta.annotation.PostConstruct;
+import jakarta.enterprise.event.Event;
+import jakarta.faces.application.FacesMessage;
+import jakarta.faces.context.FacesContext;
 import jakarta.faces.view.ViewScoped;
 import jakarta.inject.Inject;
 import jakarta.inject.Named;
@@ -21,9 +24,12 @@ import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.PropertyResourceBundle;
 
 /**
  * Backing bean for the scientific forum page.
+ *
+ * @author Stefanie Gürster
  */
 @ViewScoped
 @Named
@@ -49,6 +55,12 @@ public class ScientificForumBacking implements Serializable {
 
     @Inject
     private ConfigPropagator config;
+
+    @Inject
+    private Event<UIMessage> uiMessageEvent;
+
+    @Inject
+    private transient PropertyResourceBundle resourceBundle;
 
     private ScientificForum forum;
 
@@ -277,9 +289,15 @@ public class ScientificForumBacking implements Serializable {
     /**
      * Remove a specific user form the list of editors.
      */
-    public void removeEditor(User editor) {
+    public String removeEditor(User editor) {
+        if (editors.size() <= 1) {
+            uiMessageEvent.fire(new UIMessage(resourceBundle.getString("contactAdmin"), MessageCategory.WARNING));
+            FacesContext.getCurrentInstance().getExternalContext().getFlash().setKeepMessages(true);
+            return "/views/authenticated/scientificForum.xhtml?faces-redirect=true&id=" + forum.getId();
+        }
         forumService.removeEditor(editor, forum);
         editors.remove(editor);
+        return "/views/authenticated/scientificForum.xhtml?faces-redirect=true&id=" + forum.getId();
     }
 
     /**
